v2/requirements.txt
elasticsearch_dsl==7.4.1
minio==7.1.13
pandas==1.5.3
psutil==5.9.4
pytest==7.2.2
redis==4.5.1
requests==2.27.1
openpyxl==3.1.2
python_dotenv==1.0.1
python-slugify==8.0.4